chore(refactor): remove dedicate start, end, extension components
