+++ src/component/Gateway.tsx
@@ -2,7 +2,10 @@ import * as React from "react";
 
 import { HorizontalStroke } from "./HorizontalStroke";
 
-export const Gateway: React.FunctionComponent<{ type: "converging" } | { type: "diverging"; children?: React.ReactNode }> = ({ type, children }) => (
+export const Gateway: React.FunctionComponent<{ type: "converging"; children?: never } | { type: "diverging"; children?: React.ReactNode }> = ({
+    type,
+    children
+}) => (
     <>
         <div className="arrow" />
         <div className={`flow-element gateway-element ${type}`} />
