--- src/component/flow-element/End.tsx
@@ -1,8 +1,0 @@
-import * as React from "react";
-
-export const End: React.FunctionComponent = () => (
-    <>
-        <div className="arrow" />
-        <div className="flow-element end-element" />
-    </>
-);
