--- src/component/flow-element/Start.tsx
@@ -1,3 +1,0 @@
-import * as React from "react";
-
-export const Start: React.FunctionComponent = () => <div className="flow-element start-element" />;
