--- src/component/flow-element/StrokeExtension.tsx
@@ -1,3 +1,0 @@
-import * as React from "react";
-
-export const StrokeExtension: React.FunctionComponent = () => <div className="stroke-horizontal" />;
