--- src/component/flow-element/index.ts
@@ -1,6 +1,0 @@
-export { Start } from "./Start";
-export { ContentElement } from "./ContentElement";
-export { Gateway } from "./Gateway";
-export { HorizontalStroke } from "./HorizontalStroke";
-export { StrokeExtension } from "./StrokeExtension";
-export { End } from "./End";
