+++ test/component/ContentElement.test.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 
-import { Start } from "../../../src/component/flow-element/Start";
+import { ContentElement } from "../../src/component/ContentElement";
 
 describe("renders correctly", () => {
     it("without props", () => {
-        const component = shallow(<Start />);
+        const component = shallow(<ContentElement>text</ContentElement>);
         expect(component).toMatchSnapshot();
     });
 });
