+++ test/component/Gateway.test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 
-import { Gateway } from "../../../src/component/flow-element/Gateway";
+import { Gateway } from "../../src/component/Gateway";
 
 describe("renders correctly", () => {
     it("as converging gateway", () => {
